Fix fix buffer length fix
Searching in string must not have null characters in the beginning of the string. Searching in reverse logic has buffer align at the end instead at the beginning when parsing objects. Helps to reset the buffer to something else than null char, if we read less text than buffer size.